+++ src/main/kotlin/com/mituuz/fuzzier/entities/RowContainer.kt
@@ -29,13 +29,53 @@ class RowContainer(
     filePath: String,
     basePath: String,
     filename: String,
-    val rowNumber: Int
+    val rowNumber: Int,
+    val columnNumber: Int,
+    val trimmedRow: String
 ) : FuzzyContainer(filePath, basePath, filename) {
+    companion object {
+        /**
+         * Create a row container from a string
+         * <br><br>
+         * <h2>ripgrep</h2>
+         * ```
+         * ./src/main/kotlin/com/mituuz/fuzzier/components/TestBenchComponent.kt:205:33:            moduleFileIndex.iterateContent(contentIterator)
+         * ```
+         * <h2>grep</h2>
+         * ```
+         * ./src/main/kotlin/com/mituuz/fuzzier/components/TestBenchComponent.kt:205:            moduleFileIndex.iterateContent(contentIterator)
+         * ```
+         * <h2>findstr</h2>
+         * ```
+         * src\main\kotlin\com\mituuz\fuzzier\components\TestBenchComponent.kt:205:            moduleFileIndex.iterateContent(contentIterator)
+         * ```
+         */
+        fun rowContainerFromString(row: String, basePath: String, isRg: Boolean, isWindows: Boolean): RowContainer {
+            val parts = row.split(":")
+            var filePath = parts[0].removePrefix(".")
+            val filename = filePath.substringAfterLast(if (isWindows) "\\" else "/")
+            val rowNumber = parts[1].toInt()
+            var columnNumber: Int
+            var trimmedRow: String
+            if (isRg) {
+                columnNumber = parts[2].toInt()
+                trimmedRow = parts[3]
+            } else {
+                if (isWindows) {
+                    filePath = "/$filePath"
+                }
+                columnNumber = 0
+                trimmedRow = parts[2]
+            }
+            return RowContainer(filePath, basePath, filename, rowNumber, columnNumber, trimmedRow)
+        }
+    }
+
     override fun getDisplayString(state: FuzzierGlobalSettingsService.State): String {
         if (state.filenameType == FilenameType.DEBUG) {
             return toString()
         }
-        return "$filename:$rowNumber"
+        return "$filename $rowNumber:$columnNumber:$trimmedRow"
     }
 
     override fun toString(): String {